Create deps.edn file
Start transitioning the project.clj to deps.edn. Some extra work to run
the tests through Kaocha instead of Midje was made, because I couldn't
find how to reliably run the tests with Midje. This also uncovered a
problem when running the tests with clojure.tools.logging, some tests
were not passing.